Split extension tests
